Moved the signlr logic to a service
