+++ CastIt.Android/lib/application/settings/settings_bloc.dart
@@ -7,11 +7,10 @@ import 'package:castit/domain/enums/subtitle_bg_color_type.dart';
 import 'package:castit/domain/enums/text_track_font_generic_family_type.dart';
 import 'package:castit/domain/extensions/string_extensions.dart';
 import 'package:castit/domain/models/models.dart';
+import 'package:castit/domain/services/castit_hub_client_service.dart';
 import 'package:castit/domain/services/settings_service.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:package_info_plus/package_info_plus.dart';
-
-import '../server_ws/server_ws_bloc.dart';
 
 part 'settings_bloc.freezed.dart';
 part 'settings_event.dart';
@@ -19,12 +18,12 @@ part 'settings_state.dart';
 
 class SettingsBloc extends Bloc<SettingsEvent, SettingsState> {
   final SettingsService _settings;
-  final ServerWsBloc _serverWsBloc;
   final MainBloc _mainBloc;
+  final CastItHubClientService _castItHub;
 
   SettingsState get initialState => SettingsState.loading();
 
-  SettingsBloc(this._settings, this._serverWsBloc, this._mainBloc) : super(SettingsState.loading());
+  SettingsBloc(this._settings, this._mainBloc, this._castItHub) : super(SettingsState.loading());
 
   _LoadedState get currentState => state as _LoadedState;
 
@@ -109,15 +108,15 @@ class SettingsBloc extends Bloc<SettingsEvent, SettingsState> {
   }
 
   void listenHubEvents() {
-    _serverWsBloc.connected.stream.listen((event) {
+    _castItHub.connected.stream.listen((event) {
       add(SettingsEvent.load());
     });
 
-    _serverWsBloc.settingsChanged.stream.listen((settings) {
+    _castItHub.settingsChanged.stream.listen((settings) {
       add(SettingsEvent.connected(settings: settings));
     });
 
-    _serverWsBloc.disconnected.stream.listen((_) {
+    _castItHub.disconnected.stream.listen((_) {
       add(SettingsEvent.disconnected());
     });
   }
